requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-core==1.36.0
azure-functions==1.24.0
azure-servicebus==7.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
uamqp
websocket-client
azure-cosmos
azure-functions
azure-cosmos
azure-storage-blob
pandas
python-dateutil
azure-storage-queue


